Completely removed everything from old boardview into new GameView. Dice animation now works fine!
